Initial checkin of some utilities to assist command line apps
This is basically turning the useful bits of the progress example into
a little module for Ilan's benefit.
