+++ encore/terminal/api.py
@@ -1,0 +1,10 @@
+#
+# (C) Copyright 2011 Enthought, Inc., Austin, TX
+# All right reserved.
+#
+# This file is open source software distributed according to the terms in LICENSE.txt
+#
+
+# Local imports
+from .progress_display import ProgressDisplay, SpinWriter, ProgressWriter
+from .utils import write_static
